Factor out a waitFor function: It's not just resources which can be waited for
