+++ operator-common/src/main/java/io/strimzi/operator/common/operator/resource/AbstractResourceOperator.java
@@ -10,9 +10,9 @@ import io.fabric8.kubernetes.client.KubernetesClient;
 import io.fabric8.kubernetes.client.dsl.MixedOperation;
 import io.fabric8.kubernetes.client.dsl.NonNamespaceOperation;
 import io.fabric8.kubernetes.client.dsl.Resource;
+import io.strimzi.operator.common.Util;
 import io.strimzi.operator.common.model.Labels;
 import io.vertx.core.Future;
-import io.vertx.core.Handler;
 import io.vertx.core.Vertx;
 import org.apache.logging.log4j.LogManager;
 import org.apache.logging.log4j.Logger;
@@ -227,50 +227,10 @@ public abstract class AbstractResourceOperator<C extends KubernetesClient, T ext
      * @param predicate The predicate.
      */
     public Future<Void> waitFor(String namespace, String name, long pollIntervalMs, final long timeoutMs, BiPredicate<String, String> predicate) {
-        Future<Void> fut = Future.future();
-        log.debug("Waiting for {} resource {} in namespace {} to get ready", resourceKind, name, namespace);
-        long deadline = System.currentTimeMillis() + timeoutMs;
-        Handler<Long> handler = new Handler<Long>() {
-            @Override
-            public void handle(Long timerId) {
-                vertx.createSharedWorkerExecutor("kubernetes-ops-pool").executeBlocking(
-                    future -> {
-                        try {
-                            if (predicate.test(namespace, name))   {
-                                future.complete();
-                            } else {
-                                log.trace("{} {} in namespace {} is not ready", resourceKind, name, namespace);
-                                future.fail("Not ready yet");
-                            }
-                        } catch (Throwable e) {
-                            log.warn("Caught exception while waiting for {} {} in namespace {} to get ready", resourceKind, name, namespace, e);
-                            future.fail(e);
-                        }
-                    },
-                    true,
-                    res -> {
-                        if (res.succeeded()) {
-                            log.debug("{} {} in namespace {} is ready", resourceKind, name, namespace);
-                            fut.complete();
-                        } else {
-                            long timeLeft = deadline - System.currentTimeMillis();
-                            if (timeLeft <= 0) {
-                                String exceptionMessage = String.format("Exceeded timeout of %dms while waiting for %s %s in namespace %s to be ready", timeoutMs, resourceKind, name, namespace);
-                                log.error(exceptionMessage);
-                                fut.fail(new TimeoutException(exceptionMessage));
-                            } else {
-                                // Schedule ourselves to run again
-                                vertx.setTimer(Math.min(pollIntervalMs, timeLeft), this);
-                            }
-                        }
-                    }
-                );
-            }
-        };
-
-        // Call the handler ourselves the first time
-        handler.handle(null);
-
-        return fut;
+        return Util.waitFor(vertx,
+            String.format("%s resource %s in namespace %s", resourceKind, name, namespace),
+            pollIntervalMs,
+            timeoutMs,
+            () -> predicate.test(namespace, name));
     }
 }
